[New annotator] Merge presence bloc and label list bloc in a unique Labels bloc Update labels shortcut to enable new box with the focus label without changing current box label
