+++ frontend/src/view/annotator/Annotator.tsx
@@ -14,8 +14,6 @@ import { SpectrogramRender } from "@/view/annotator/tools/spectrogram/Spectrogra
 import { SpectrogramDownloadButton } from "@/view/annotator/tools/buttons/SpectrogramDownload.tsx";
 import { PlayPauseButton } from "@/view/annotator/tools/buttons/PlayPause.tsx";
 import { NavigationButtons } from "@/view/annotator/tools/buttons/Navigation.tsx";
-import { PresenceAbsence } from "@/view/annotator/tools/bloc/PresenceAbsence.tsx";
-import { LabelList } from "@/view/annotator/tools/bloc/LabelList.tsx";
 import { CurrentAnnotation } from "@/view/annotator/tools/bloc/CurrentAnnotation.tsx";
 import { Comment } from "@/view/annotator/tools/bloc/Comment.tsx";
 import { ConfidenceIndicator } from "@/view/annotator/tools/bloc/ConfidenceIndicator.tsx";
@@ -24,6 +22,7 @@ import { PlaybackRateSelect } from "@/view/annotator/tools/select/PlaybackRate.t
 import { useToast } from "@/service/ui";
 import { useAnnotator } from "@/service/annotator/hook.ts";
 import { useFileDuration } from '@/service/annotator/spectrogram';
+import { Labels } from "@/view/annotator/tools/bloc/Labels.tsx";
 
 export const Annotator: React.FC = () => {
   const {
@@ -115,11 +114,10 @@ export const Annotator: React.FC = () => {
         <div className={ [styles.blocContainer, campaign?.usage === 'Check' ? styles.check : styles.create ].join(' ') }>
           { annotatorData?.is_assigned && campaign?.usage === 'Create' && <Fragment>
               <Comment/>
-              <PresenceAbsence/>
-              <LabelList/>
+              <Labels/>
+              <ConfidenceIndicator/>
               <CurrentAnnotation/>
               <Results/>
-              <ConfidenceIndicator/>
           </Fragment> }
           { annotatorData?.is_assigned && campaign?.usage === 'Check' && <Fragment>
               <CurrentAnnotation/>
