+++ frontend/src/view/annotator/annotator.module.scss
@@ -107,6 +107,10 @@
 
     &.create {
       grid-template-columns: 32ch 2fr 1fr;
+
+      > *:last-child {
+        grid-column: 2 / -1;
+      }
     }
 
     &.check {
